fix: per-100 energy cost, last-done gap, form label, overflow headroom
Four small items from the re-review of 22ba592 (new commit, that one left
untouched):

1. The Energy section's cost row now reads "Energy cost per 100 {unit}"
   (energy.cost-per-100), computed as cost_per_counter_milli * 100, matching
   the app's existing per-100 convention (insights.consumption). A single-unit
   rate rounded to cents was both uninformative and ~40% high (600 milli ->
   "€0.01" instead of the true "€0.60" per 100 km). TripTotals' own "Energy
   cost" row and the trip row's "≈ €1.20" are untouched -- those are already
   correct absolute costs over real distances, not per-unit rates.
2. last_done now excludes `fuel` only when the object has a fuel_unit set:
   `fuel` is offered on car/e_bike/motorcycle regardless, and an object that
   logs fills but has no fuel_unit gets no Energy section either, so
   "distance since the last fill" still belongs there. Two tests: fuel_unit
   set -> excluded (renamed from the old blanket-exclusion test), no
   fuel_unit -> a repeated fill still counts.
3. ActivityForm's Amount label no longer invents "l"/"gal" when the object
   has no fuel_unit -- matches the bare number Timeline.svelte already shows
   for the same entries.
4. domain::energy's two unbounded multiplications (distance * 1_000_000, and
   quantity * price in the cost fallback) now widen to i128 before dividing
   back down to i64, so a fat-fingered import's huge counter/quantity/price
   can't panic a debug build. Every existing expectation unchanged.

+++ src/domain/energy.rs
@@ -115,7 +115,12 @@ pub fn energy(charges: &[Charge], trips: &[Trip], price_milli: Option<i64>) -> E
             // (An earlier version divided by only 1000, which cancelled `quantity_milli`'s own
             // x1000 instead of compounding it, landing 1000x too small -- e.g. reading 0.1
             // km/kWh for a rate that is really 50 times higher.)
-            Some(distance * 1_000_000 / quantity)
+            //
+            // `i128`: `counter_value` is only validated `>= 0`, not bounded above, so
+            // `distance * 1_000_000` could in principle overflow `i64` for a fat-fingered import
+            // -- widen for the multiply and narrow back once the division has brought the
+            // result down to a sane range, rather than risk a panic in a debug build.
+            Some((distance as i128 * 1_000_000 / quantity as i128) as i64)
         })
         .collect();
     let distance_per_unit_milli = mean(&unit_rates);
@@ -137,7 +142,10 @@ pub fn energy(charges: &[Charge], trips: &[Trip], price_milli: Option<i64>) -> E
                 None => {
                     let quantity = b.quantity_milli?;
                     let price = price_milli?;
-                    quantity * price / 1000 / distance
+                    // `i128` for the same reason as the unit-rate multiply above: `quantity` and
+                    // `price` are each only validated `>= 0`, and their product alone (before
+                    // this is even divided down) could overflow `i64`.
+                    (quantity as i128 * price as i128 / 1000 / distance as i128) as i64
                 },
             };
             Some(rate)
